Increase expressiveness in API types 1/2
Looking through the spring web handlers in io.spring.api one notices lots of `ResponseEntity` and `ResponseEntity<?>` references. These do not add information to the API implementation as spring boot wraps any non-ResponseEntity type comming out of a web handler (public @RestController methods annotated with @XxxMapping) with a `ResponseEntity.ok` instance.
In order to remove unecessary code and increase the readability of the source code I removed all references to `ResponseEntity.ok()` and forced actual types on the return values of the API classes. We can do this since all integration tests (postman collection) test for the correct response headers which means am uncatched change in the spring boot defaults would be immediately noticed in CI.
This is the first part of two commits intended to increase the readability and succinctness of the Api implementation. The next commit will introduce a replacement to the multiple `Map<k,v>` returns that currently taint the Api classes.