Increase expressiveness in API types 2/2
Example implementation of a factory for responses that cleans up the handler code by removing the imperative data wrangling in favor of a declarative style. This is just an example, stub, implementation in Articles and Comments and could be extended to the whole codebase. The current Response Factory uses method overloading to implement the business rules that are required to produce the correct responses. It can be possible that a more involved solution will be required if this is to be rolled out to the whole API.